% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{add_matrix_metadata}
\alias{add_matrix_metadata}
\title{Add metadata to matrix for ComplexHeatmap}
\usage{
add_matrix_metadata(m, gr, target_name, signal_name)
}
\arguments{
\item{m}{The matrix to update.}

\item{gr}{A GRanges corresponding to the regions to subset the coverages.
All the regions must have the same width (see
?GenomicRanges::resize to resize regions).}

\item{target_name}{The name of the center region of the matrix. Will be
shown at the bottom of the matrix.}

\item{signal_name}{The name of the matrix. Will be shown on top of the
matrix.}
}
\value{
A normalizedMatrix ready to be used with ComplexHeatmap.
}
\description{
Add metadata to matrix for ComplexHeatmap
}
\examples{
filenames <- get_demo_bedGraph_files()
cov <- import_bedgraphs(filenames)

gr <- import_peaks(get_demo_peaks_file())
m <- coverage_2_matrix(cov[[1]], gr)
m <- add_matrix_metadata(m, gr, "TSS", "Peaks")
}
