% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{coverage_2_matrix}
\alias{coverage_2_matrix}
\title{Convert coverages to matrix}
\usage{
coverage_2_matrix(coverage, gr, ncol = 100)
}
\arguments{
\item{coverage}{A single coverage obtained with the import_bedgraphs
function.}

\item{gr}{A GRanges corresponding to the subset of regions to subset
the coverages. All the regions must have the same width (see
?GenomicRanges::resize to resize regions).}

\item{ncol}{The number of columns in the result matrix. (Default = 100).}
}
\value{
A matrix with ncol columns and length(gr) rows.
}
\description{
Convert coverages to matrix
}
\examples{
filenames <- get_demo_bedGraph_files()
cov <- import_bedgraphs(filenames)

gr <- import_peaks(get_demo_peaks_file())
m <- coverage_2_matrix(cov[[1]], gr)

}
