% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_bedGraph_files}
\alias{get_demo_bedGraph_files}
\title{Get bedGraph filenames for demo}
\usage{
get_demo_bedGraph_files()
}
\value{
A vector with the bedGraph filenames used for demo
}
\description{
Get bedGraph filenames for demo
}
\examples{
bg_file <- get_demo_bedGraph_files()

}
