% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_peaks_file}
\alias{get_demo_peaks_file}
\title{Get narrowPeak filename for demo}
\usage{
get_demo_peaks_file()
}
\value{
The complete path to a demo narrowPeak file
}
\description{
Get narrowPeak filename for demo
}
\examples{
peak_file <- get_demo_peaks_file()

}
