% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_peaks}
\alias{import_peaks}
\title{Import peak files}
\usage{
import_peaks(filename, genome = NULL, keep_standard_chromosomes = TRUE)
}
\arguments{
\item{filename}{Path to the peak file.}

\item{genome}{Add seqinfo from genome (see ?fetchExtendedChromInfoFromUCSC
for the list of available genomes). Value must be NULL or a character
string (i.e.: "hg38").}

\item{keep_standard_chromosomes}{Remove alternative chromosomes? (Default: TRUE)}
}
\value{
A GRanges corresponding the regions in the peak file.
}
\description{
Only works for .narrowPeak or .broadPeak file formats.
}
\examples{
peaks_file <- get_demo_peaks_file()
peaks <- import_peaks(peaks_file)

}
