% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hm_utilities.R
\name{intoNbins}
\alias{intoNbins}
\title{Split a GRanges into N bins}
\usage{
intoNbins(gr, n = 10)
}
\arguments{
\item{n}{Number of bins to produce.}
}
\value{
A GRanges object splitted into N bins
}
\description{
Originally posted by Martin Morgan:
https://stat.ethz.ch/pipermail/bioconductor/2012-September/047923.html
}
\details{
Wparam gr A GRanges with only one seqnames value.
}
\examples{
  gr <- GRanges("chr1", IRanges(c(100, 300), c(200, 500))
  gr <- intoNbins(gr)
}
