% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_bedgraphs}
\alias{import_bedgraphs}
\title{Import bedgraph files}
\usage{
import_bedgraphs(filenames, filter_negative_coverages = TRUE,
  keep_standard_chromosomes = TRUE)
}
\arguments{
\item{filenames}{Paths to the bedGraph files.}

\item{filter_negative_coverages}{Convert negative coverage values to 0? (Default: TRUE)}

\item{keep_standard_chromosomes}{Remove alternative chromosomes? (Default: TRUE)}
}
\value{
A list of GRanges (one element per file).
}
\description{
Import bedgraph files
}
\examples{
\dontrun{
    filenames <- c("file1.bg", "file2.bg")
    cov <- import_bedgraphs(filenames)
}

}
