% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_peaks}
\alias{import_peaks}
\title{Import peak files}
\usage{
import_peaks(filenames)
}
\arguments{
\item{filenames}{Paths to the peak files.}
}
\value{
A list of GRanges (one element per file).
}
\description{
Only works for .narrowPeak or .broadPeak file formats.
}
\examples{
peaks_file <- get_demo_peaks_file()
peaks <- import_peaks(peaks_file)

}
